backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
transformers==4.33.2
chromadb==0.4.8
qdrant-client==1.5.4
python-dotenv==1.0.0
sqlalchemy==2.0.21
alembic==1.12.0
pytest==7.4.2
black==23.7.0
flake8==6.0.0
pymongo==4.6.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0
